requirements.txt
streamlit
boto3
pillow

